redesigned the type system
