+++ emc.vcxproj
@@ -116,6 +116,8 @@
     <ClInclude Include="src\linker.h" />
     <ClInclude Include="src\llvm.h" />
     <ClInclude Include="src\parser.h" />
+    <ClInclude Include="src\symbols.h" />
+    <ClInclude Include="src\types.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
